rename CustomTrim to TrimKey and rename package
